tech?(api): move handle-stage-acquisition in evaluation bounded context
