+++ api/lib/domain/usecases/index.js
@@ -51,7 +51,6 @@ import * as competenceRepository from '../../../src/shared/infrastructure/reposi
 import * as courseRepository from '../../../src/shared/infrastructure/repositories/course-repository.js';
 import * as frameworkRepository from '../../../src/shared/infrastructure/repositories/framework-repository.js';
 import { repositories as sharedRepositories } from '../../../src/shared/infrastructure/repositories/index.js';
-import * as knowledgeElementRepository from '../../../src/shared/infrastructure/repositories/knowledge-element-repository.js';
 import * as organizationLearnerRepository from '../../../src/shared/infrastructure/repositories/organization-learner-repository.js';
 import * as organizationRepository from '../../../src/shared/infrastructure/repositories/organization-repository.js';
 import * as skillRepository from '../../../src/shared/infrastructure/repositories/skill-repository.js';
@@ -181,7 +180,6 @@ const path = dirname(fileURLToPath(import.meta.url));
 
 const usecasesWithoutInjectedDependencies = {
   ...(await importNamedExportsFromDirectory({ path: join(path, './'), ignoredFileNames: ['index.js'] })),
-  ...(await importNamedExportsFromDirectory({ path: join(path, './stages') })),
 };
 
 const usecases = injectDependencies(usecasesWithoutInjectedDependencies, dependencies);
